+++ include/libopencm3/stm32/f1/i2c.h
@@ -31,7 +31,7 @@ LGPL License Terms @ref lgpl_license
 #ifndef LIBOPENCM3_I2C_H
 #define LIBOPENCM3_I2C_H
 
-#include <libopencm3/stm32/common/i2c_common_all.h>
+#include <libopencm3/stm32/common/i2c_common_v1.h>
 
 #endif
