+++ include/libopencm3/stm32/f2/i2c.h
@@ -31,7 +31,13 @@ LGPL License Terms @ref lgpl_license
 #ifndef LIBOPENCM3_I2C_H
 #define LIBOPENCM3_I2C_H
 
-#include <libopencm3/stm32/common/i2c_common_f24.h>
+#include <libopencm3/stm32/common/i2c_common_v1.h>
+
+/**@{*/
+
+#define I2C3				I2C3_BASE
+
+/**@}*/
 
 #endif
