+++ include/libopencm3/stm32/i2c.h
@@ -30,8 +30,14 @@
 #       include <libopencm3/stm32/f3/i2c.h>
 #elif defined(STM32F4)
 #       include <libopencm3/stm32/f4/i2c.h>
+#elif defined(STM32F7)
+#       include <libopencm3/stm32/f7/i2c.h>
+#elif defined(STM32L0)
+#       include <libopencm3/stm32/l0/i2c.h>
 #elif defined(STM32L1)
 #       include <libopencm3/stm32/l1/i2c.h>
+#elif defined(STM32L4)
+#       include <libopencm3/stm32/l4/i2c.h>
 #else
 #       error "stm32 family not defined."
 #endif
