+++ lib/stm32/f0/Makefile
@@ -36,7 +36,7 @@ TGT_CFLAGS      += $(DEBUG_FLAGS)
 ARFLAGS		= rcs
 
 OBJS		= flash.o rcc.o usart.o dma.o rtc.o comparator.o crc.o \
-                  dac.o i2c.o iwdg.o pwr.o gpio.o timer.o adc.o desig.o
+                  dac.o iwdg.o pwr.o gpio.o timer.o adc.o desig.o
 
 OBJS            += gpio_common_all.o gpio_common_f0234.o crc_common_all.o \
                    pwr_common_all.o iwdg_common_all.o rtc_common_l1f024.o \
@@ -46,6 +46,7 @@ OBJS            += gpio_common_all.o gpio_common_f0234.o crc_common_all.o \
 OBJS		+= adc_common_v2.o
 OBJS		+= crs_common_all.o
 OBJS		+= usart_common_v2.o
+OBJS		+= i2c_common_v2.o
 
 OBJS		+= usb.o usb_control.o usb_standard.o
 OBJS		+= st_usbfs_core.o st_usbfs_v2.o
