+++ lib/stm32/l1/Makefile
@@ -38,7 +38,7 @@ OBJS		= crc.o desig.o flash.o rcc.o usart.o dma.o lcd.o
 OBJS		+= crc_common_all.o dac_common_all.o
 OBJS		+= dma_common_l1f013.o
 OBJS		+= gpio_common_all.o gpio_common_f0234.o
-OBJS		+= i2c_common_all.o iwdg_common_all.o
+OBJS		+= i2c_common_v1.o iwdg_common_all.o
 OBJS		+= pwr_common_all.o pwr_common_l01.o rtc_common_l1f024.o
 OBJS		+= spi_common_all.o spi_common_l1f124.o timer_common_all.o
 OBJS		+= usart_common_all.o usart_common_f124.o
